Added an optional requirement to README
Since as an install the files get compiled, settings can't be read by external python scripts, gadis-remote being one such script. it needs to temporarily decompile and read the settings to find the remote port to connect to (default host is localhost, this can be changed before compile time on gadis-remote.py). It can however be optional by simply modifying gadis-remote,py before compile time. (Line 13 as of this commit)